// asset-management/src/generated/models/Cr240_assetassignmentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr240_assetassignmentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr240_assetassignmentsstatecode = keyof typeof Cr240_assetassignmentsstatecode;
export const Cr240_assetassignmentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr240_assetassignmentsstatuscode = keyof typeof Cr240_assetassignmentsstatuscode;

export interface Cr240_assetassignmentsBase {
  cr240_assetassignmentid: string;
  "cr240_AssetID@odata.bind"?: string;
  cr240_assignmentdate?: string;
  cr240_assignmentname?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr240_assetassignmentsstatecode;
  statuscode?: Cr240_assetassignmentsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr240_assetassignments extends Cr240_assetassignmentsBase {
  cr240_assetidname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr240_assetid?: object;
  _cr240_assetid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}